+++ src/components/frame/Extensions/layouts/index.ts
@@ -1,2 +1,3 @@
 export { default as ColonyLayout } from './ColonyLayout';
+export { default as SharedLayout } from './SharedLayout';
 export { default as UserNavigationWrapper } from './UserNavigationWrapper';
