+++ src/components/frame/Extensions/layouts/types.ts
@@ -1,6 +1,13 @@
+import { NavigationSidebarProps } from '~v5/frame/NavigationSidebar/types';
 import type { CalamityBannerItemProps } from '~v5/shared/CalamityBanner/types';
 
 export interface UseCalamityBannerInfoReturnType {
   canUpgrade?: boolean;
   calamityBannerItems: CalamityBannerItemProps[];
+}
+
+export interface SharedLayoutProps {
+  mobileBottomContent?: NavigationSidebarProps['mobileBottomContent'];
+  hamburgerLabel?: NavigationSidebarProps['hamburgerLabel'];
+  mainMenuItems?: NavigationSidebarProps['mainMenuItems'];
 }
