+++ src/components/v5/frame/PageLayout/partials/PageHeader/types.ts
@@ -1,6 +1,7 @@
 import { PageHeadingProps } from '../PageHeading/types';
 
-export interface PageHeaderProps extends Omit<PageHeadingProps, 'className'> {
+export interface PageHeaderProps {
   userNavigation: React.ReactNode;
+  pageHeadingProps?: PageHeadingProps;
   className?: string;
 }
