+++ src/stories/frame/PageLayout.stories.tsx
@@ -25,24 +25,26 @@ const pageLayoutMeta: Meta<typeof PageLayout> = {
   args: {
     children: <div className="h-screen bg-blue-200">body</div>,
     headerProps: {
-      title: 'Members',
-      breadcrumbs: [
-        {
-          key: '1',
-          label: 'Metacolony',
-          href: '/',
-        },
-        {
-          key: '2',
-          dropdownOptions: [
-            {
-              label: 'All teams',
-              href: '/all-teams',
-            },
-          ],
-          selectedValue: '/all-teams',
-        },
-      ],
+      pageHeadingProps: {
+        title: 'Members',
+        breadcrumbs: [
+          {
+            key: '1',
+            label: 'Metacolony',
+            href: '/',
+          },
+          {
+            key: '2',
+            dropdownOptions: [
+              {
+                label: 'All teams',
+                href: '/all-teams',
+              },
+            ],
+            selectedValue: '/all-teams',
+          },
+        ],
+      },
       userNavigation: <p>user navigation</p>,
     },
     navigationSidebarProps: {
